Update Sumupto_0_ofdistinct_array_integer.py
Write a function that given an integer N(1<=N<=100) returns  an array containing N distinct Integers that sum up to 0. The function can return any such array. For example, given N=4, the function could return (1,0,-3, 2], and for N=3 one of the possible answers is [-1,0,1] (but there are many more correct answers).